+++ tests/testthat/test-hit_api.R
@@ -17,7 +17,7 @@ test_that("hit_api_error_offline",{
    # bad path and too complex shape
    req <- build_req(path = "api/rpg/v1",
                     annee = 2013,
-                    geom = shp_to_geojson(st_buffer(point, 10)))
+                    geom = shp_to_geojson(sf::st_buffer(point, 100)))
    expect_error(hit_api(req), "overly complex shape")
    expect_error(hit_api(NA), "`req` must be an HTTP request")
 })